requirements.txt
algoliasearch==2.6.3
algoliasearch-django==2.0.0
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.0.10
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.6
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
